+++ src/Auth/useAuthState.ts
@@ -1,5 +1,6 @@
-import { AuthContext, AuthContextState } from './../../contexts/AuthContext';
+import { AuthContext } from './AuthContext';
 import { useContext } from 'react';
+import { AuthContextState } from './AuthContext.types';
 
 const useAuthState = (): AuthContextState => {
   const { state } = useContext(AuthContext);
