--- src/hooks/Auth/useAuthActions.types.ts
@@ -1,4 +1,0 @@
-export interface LoginInfo {
-  identifier: string;
-  password: string;
-}
